+++ http/http_client_jetty/src/main/kotlin/com/hexagonkt/http/client/jetty/JettyClientAdapter.kt
@@ -2,7 +2,7 @@ package com.hexagonkt.http.client.jetty
 
 import com.hexagonkt.core.media.TEXT_EVENT_STREAM
 import com.hexagonkt.core.security.loadKeyStore
-import com.hexagonkt.http.bodyToBytes
+import com.hexagonkt.http.handlers.bodyToBytes
 import com.hexagonkt.http.CHECKED_HEADERS
 import com.hexagonkt.http.client.HttpClient
 import com.hexagonkt.http.client.HttpClientPort
