+++ http/http_server_servlet/src/main/kotlin/com/hexagonkt/http/server/servlet/ServletFilter.kt
@@ -3,7 +3,7 @@ package com.hexagonkt.http.server.servlet
 import com.hexagonkt.core.logging.Logger
 import com.hexagonkt.core.media.TEXT_PLAIN
 import com.hexagonkt.core.toText
-import com.hexagonkt.http.bodyToBytes
+import com.hexagonkt.http.handlers.bodyToBytes
 import com.hexagonkt.http.handlers.HttpHandler
 import com.hexagonkt.http.model.HttpResponse
 import com.hexagonkt.http.model.HttpResponsePort
